requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.3.0
faiss-cpu
python-dotenv
pypdf
docx2txt
numpy
